# openapi.yaml
openapi: 3.1.0
info:
  title: DCF Automated Intake Tool API
  description: |
    Demo platform API for Massachusetts DCF AIT.
    Contract-first design — see Docs/DCF-AIT-PLATFORM-ARCHITECTURE.md and openspec/specs/ (8 domains).
  version: 1.0.0
  contact:
    name: InApp HHS Practice
servers:
  - url: https://localhost/api/v1
    description: Local Docker (via nginx TLS)
  - url: http://localhost:8080/api/v1
    description: Direct API (development)
...
security:
  - bearerAuth: []
paths:
...
  /auth/demo-login:
    post:
      tags: [Auth]
      summary: Demo role login (mockup role picker)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DemoLoginRequest'
      responses:
        '200':
          description: JWT issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /cases:
    get:
      tags: [Cases]
      summary: List cases for current role
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [in_progress, pending_review, assigned, closed]
        - name: emergency
          in: query
          schema:
            type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CaseSummary'
    post:
      tags: [Cases]
      summary: Create new 51A case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCaseRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseDetail'
  /cases/{caseId}:
    get:
      tags: [Cases]
      summary: Get case detail (RBAC applies; admin receives redacted)
      parameters:
        - $ref: '#/components/parameters/caseId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseDetail'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Cases]
      summary: Update case metadata or 51A fields
      parameters:
        - $ref: '#/components/parameters/caseId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCaseRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseDetail'
  /cases/{caseId}/audio:
    post:
      tags: [Intake]
      summary: Upload call recording (multipart)
      parameters:
        - $ref: '#/components/parameters/caseId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '202':
          description: Upload accepted; pipeline started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineStatus'
...
  /cases/{caseId}/submit:
    post:
      tags: [Intake]
      summary: Screener submit to supervisor (requires 51A form checkpoint complete)
      parameters:
        - $ref: '#/components/parameters/caseId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseDetail'
        '409':
          description: 51A form checkpoint not complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Form51AValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    caseId:
      name: caseId
      in: path
      required: true
      schema:
        type: string
        example: 51A-2024-0847
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
...
    DemoLoginRequest:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [screener, supervisor, worker, admin]
        displayName:
          type: string
          example: John Smith
        areaOffice:
          type: string
          example: Springfield
    AuthTokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
        role:
          type: string
...
    CreateCaseRequest:
      type: object
      properties:
        emergencyHint:
          type: boolean
          default: false
    UpdateCaseRequest:
      type: object
      properties:
        fields:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/FieldValue'
    FieldValue:
      type: object
      properties:
        value:
          type: string
        source:
          type: string
          enum: [human, ai, ccwis]
        confirmed:
          type: boolean
    CaseSummary:
      type: object
      properties:
        caseId:
          type: string
        childDisplay:
          type: string
        status:
          type: string
        form51aCheckpointStatus:
          $ref: '#/components/schemas/Form51ACheckpointStatus'
        riskScore:
          type: integer
          minimum: 1
          maximum: 20
        triageFlagCount:
          type: integer
        emergency:
          type: boolean
        updatedAt:
          type: string
          format: date-time
    Form51ACheckpointStatus:
      type: string
      enum:
        - not_started
        - ai_populating
        - ready_for_review
        - incomplete
        - complete
        - locked
...
    Form51AValidationError:
      type: object
      properties:
        code:
          type: string
          example: FORM_51A_INCOMPLETE
        message:
          type: string
        missingFields:
          type: array
          items:
            type: string
        unconfirmedAiFields:
          type: array
          items:
            type: string
        checkpointStatus:
          $ref: '#/components/schemas/Form51ACheckpointStatus'
    CaseDetail:
      allOf:
        - $ref: '#/components/schemas/CaseSummary'
        - type: object
          properties:
            fields:
              type: object
              additionalProperties:
                $ref: '#/components/schemas/FieldValue'
            pipeline:
              $ref: '#/components/schemas/PipelineStatus'
    PipelineStatus:
      type: object
      properties:
        currentStage:
          type: string
        stages:
          type: object
          additionalProperties:
            type: string
            enum: [pending, running, complete, failed]
...